pyproject.toml
[project]
name = "researchhub"
version = "0.1.0"
description = "ResearchHub Backend"
readme = "README.md"
requires-python = ">=3.13,<3.14"
authors = [{ name = "ResearchHub", email = "dev@researchhub.com" }]
dependencies = [
    "arxiv>=1.4.8,<2.0.0",
    "atproto>=0.0.63",
    "authlib>=1.3.0,<2.0.0",
    "beautifulsoup4==4.12.2",
    "boto3==1.34.69",
    "cdp-sdk==1.40.0",
    "celery==5.6.3",
    "celery-redbeat==2.3.3",
    "channels>=4.1.0,<5.0.0",
    "channels-redis>=4.2.0,<5.0.0",
    "circle-developer-controlled-wallets>=9.2.0,<10.0.0",
    "cloudscraper>=1.2.60,<2.0.0",
    "daphne>=4.1.2,<5.0.0",
    "dj-rest-auth[with-mfa,with-social]>=7.0.2,<8.0.0",
    "django==5.2.13",
    "django-allauth>=65.13.0,<66.0.0",
    "django-celery-results==2.6.0",
    "django-cors-headers==3.5.0",
    "django-dbbackup==4.0.2",
    "django-debug-toolbar==4.2",
    "django-opensearch-dsl==0.7.0",
    "django-extensions==3.0.9",
    "django-fernet-encrypted-fields>=0.3.1",
    "django-filter>=24.3,<25.0",
    "django-health-check>=3.18.3,<4.0.0",
    "django-mjml>=1.3,<2.0",
    "django-ses>=4.7.2,<5",
    "django-silk==5.0.2",
    "django-simple-history==3.7.0",
    "django-storages==1.10.1",
    "djangorestframework==3.16.1",
    "elastic-apm==6.24.0",
    "humanize>=4.12.2,<5.0.0",
    "flower>=2.0.1,<3.0.0",
    "habanero==1.0.0",
    "lxml>=6.0.1",
    "mailchimp-marketing>=3.0.72,<4.0.0",
    "manubot==0.5.3",
    "openai>=1.45.0",
    "opensearch-py==3.0.0",
    "pandas>=2.2.3,<3.0.0",
    "pillow>=12.0.0",
    "psycopg[binary,pool]>=3.2.3,<4.0.0",
    "pyjwt>=2.12.0",
    "PyMuPDF>=1.24.13,<2.0.0",
    "python-json-logger>=4.0.0",
    "reportlab>=4.0.0",
    "pyyaml>=6.0.2",
    "q5-django-inlinecss==0.0.2",
    "requests-mock>=1.12.1,<2.0.0",
    "responses==0.21.0",
    "sentry-sdk>=2.49.0,<3.0.0",
    "stripe>=11.2.0,<12.0.0",
    "web3[tester]==7.10.0",
    "xdk>=0.4.5",
]

[tool.coverage.run]
concurrency = ["multiprocessing"]
parallel = true
omit = ["*/setup-uv-cache/*"]

[dependency-groups]
dev = [
    "black==26.3.1",
    "coverage>=7.6.1",
    "flake8==5.0.4",
    "pre-commit==2.17.0",
    "setuptools>=78.1.1",
    "tblib>=3.2.2",
]
